api/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart # parsing multipart/form-data POST requests.